requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-core
langchain-anthropic
python-dotenv
pydantic
